requirements.txt
customtkinter>=5.2
matplotlib>=3.5
pysnmp>=4.4
speedtest-cli>=2.1
manuf
psutil
scapy
ping3
numpy<2
pystray
paramiko